Add files via upload
This project implements the order four Runge-Kutta method and a predictor corrector method based on the Adams-Bashforth and Adams-Moulton schemes. The code can solve systems of up to 3 coupled non-linear first order equations. However, it should be possible to extend that to arbitrary order. The actual solvers are written with arbitrary order in mind but the practicalities of the interface mean that the class is currently limited to a maximum of 3 equations.